+++ ouzel/audio/Containers.hpp
@@ -16,7 +16,7 @@ namespace ouzel
         public:
             void reset() override;
 
-            void getSamples(uint32_t frames, uint16_t channels,
+            void getSamples(uint32_t frames, uint32_t channels,
                             uint32_t sampleRate, std::vector<float>& samples) override;
         };
 
@@ -25,7 +25,7 @@ namespace ouzel
         public:
             void reset() override;
 
-            void getSamples(uint32_t frames, uint16_t channels,
+            void getSamples(uint32_t frames, uint32_t channels,
                             uint32_t sampleRate, std::vector<float>& samples) override;
         };
 
@@ -34,7 +34,7 @@ namespace ouzel
         public:
             void reset() override;
 
-            void getSamples(uint32_t frames, uint16_t channels,
+            void getSamples(uint32_t frames, uint32_t channels,
                             uint32_t sampleRate, std::vector<float>& samples) override;
         };
     } // namespace audio
