+++ ouzel/audio/Effect.hpp
@@ -56,7 +56,7 @@ namespace ouzel
             Effect(Effect&&) = delete;
             Effect& operator=(Effect&&) = delete;
 
-            virtual void process(uint32_t frames, uint16_t channels, uint32_t sampleRate,
+            virtual void process(uint32_t frames, uint32_t channels, uint32_t sampleRate,
                                  std::vector<float>& samples) = 0;
 
             inline auto isEnabled() const noexcept { return enabled; }
