+++ ouzel/audio/Silence.cpp
@@ -18,7 +18,7 @@ namespace ouzel
             position = 0;
         }
 
-        void Silence::getSamples(uint32_t frames, uint16_t channels,
+        void Silence::getSamples(uint32_t frames, uint32_t channels,
                                  uint32_t sampleRate, std::vector<float>& samples)
         {
             std::fill(samples.begin(), samples.end(), 0.0F); // TODO: fill only the needed samples
