+++ ouzel/audio/VorbisClip.cpp
@@ -62,7 +62,7 @@ namespace ouzel
                 stb_vorbis_seek_start(vorbisStream);
             }
 
-            void getSamples(uint32_t frames, uint16_t channels, uint32_t sampleRate, std::vector<float>& samples) final
+            void getSamples(uint32_t frames, uint32_t channels, uint32_t sampleRate, std::vector<float>& samples) final
             {
             }
