+++ ouzel/audio/WavePlayer.cpp
@@ -10,7 +10,7 @@ namespace ouzel
         {
         }
 
-        void WavePlayer::getSamples(uint32_t frames, uint16_t channels,
+        void WavePlayer::getSamples(uint32_t frames, uint32_t channels,
                                     uint32_t sampleRate, std::vector<float>& samples)
         {
         }
