+++ ouzel/audio/WavePlayer.hpp
@@ -22,7 +22,7 @@ namespace ouzel
 
             void reset() override;
 
-            void getSamples(uint32_t frames, uint16_t channels,
+            void getSamples(uint32_t frames, uint32_t channels,
                             uint32_t sampleRate, std::vector<float>& samples) override;
 
         private:
