+++ ouzel/audio/mixer/Bus.hpp
@@ -34,7 +34,7 @@ namespace ouzel
                 void addInput(Source* stream);
                 void removeInput(Source* stream);
 
-                void getSamples(uint32_t frames, uint16_t channels, uint32_t sampleRate,
+                void getSamples(uint32_t frames, uint32_t channels, uint32_t sampleRate,
                                 std::vector<float>& samples) final;
 
             private:
