+++ ouzel/audio/mixer/Object.hpp
@@ -20,7 +20,7 @@ namespace ouzel
                 Object() noexcept = default;
                 virtual ~Object() = default;
 
-                virtual void getSamples(uint32_t frames, uint16_t channels, uint32_t sampleRate, std::vector<float>& samples) = 0;
+                virtual void getSamples(uint32_t frames, uint32_t channels, uint32_t sampleRate, std::vector<float>& samples) = 0;
             };
         }
     } // namespace audio
